feat: use name resolution for dot identifier notation
This PR modifies dot identifier notation so that `(.a : T)` resolves `T.a` with respect to the root namespace. This lets the notation refer to private names, follow aliases, and also take advantage of open namespaces, like for generalized field notation. The LSP completions are improved to follow how dot ident notation is resolved, but it doesn't yet take into account aliases or open namespaces.

Closes #9629