perf: avoid computing liveness twice for join point decls in RC pass
We compute the liveness information for the join point body, so the
only thing that updateJPLiveVarMap should be adding is the binding
of the params, which we can easily do ourselves.

If we supported recursive join points, I believe this would actually
be a correctness issue, but as-is it doesn't affect the output.