feat: adjust `deriving Inhabited` to use structure field defaults
This PR changes the deriving handler for `Inhabited` instances so that for structures it uses each fields' default values, if possible. For structure type `S`, the handler uses `by refine' {.. : S} <;> exact default` to construct the default value. A limitation to this is that if there are any stuck defaults due to cycles, they use `default`.

Closes #9463