feat: kernel hint for proof-by-reflection
This PR adds improved support for proof-by-reflection to the kernel
type checker. It addresses the performance issue exposed by #9854.
With this PR, whenever the kernel type-checks an argument of the form
`Eq.refl true`, it enters "proof-by-reflection" mode. In this mode,
the kernel is more eager to reduce terms. If the PR negatively
impacts existing packages, we can use a different a new `Eq.refl`
theorem as the "hint."