feat: `grind cutsat` equations in solved form
This PR ensures that equations in the `grind cutsat` module are
maintained in solved form. That is, given an equation `a*x + p = 0`
used to eliminate `x`, the linear polynomial `p` must not contain
other eliminated  variables. Before this PR, equations were maintained
in triangular form. We are now using the solved form to linearize
nonlinear terms.