feat: use reducible-and-instances when unfolding return type in elaboration
This PR modifies the app elaborator and the implicit lambda feature to reduce the return type and expected type at reducible-and-instances transparency rather than default transparency. This prevents `def` type synonyms from being unexpectedly unfolded, and it also improves performance due to less unfolding, both in elaboration and in pretty printing. **Breaking change:** `def` type synonyms that are a pi type that starts with an implicit argument may need to become `abbrev`s.