chore: eliminate uses of `intros x y z`
This PR eliminates uses of `intros x y z` (with arguments) and updates the `intros` docstring to suggest that `intro x y z` should be used instead. The `intros` tactic is historical, and can be traced all the way back to Lean 2. Since 2020, the `intro` tactic has superceded it. The `intros` tactic (without arguments) is currently still useful.