feat: pretty print sorry in "declaration uses 'sorry'"
This PR changes the "declaration uses 'sorry'" error to pretty print an actual `sorry` expression in the message. The effect is that the `sorry` is hoverable and, if it's labeled, you can "go to definition" to see where it came from.

The implementation prefers reporting synthetic sorries. These can appear even if there are no error messages if a declaration refers to a declaration that has elaboration errors. Users should focus on elaboration errors before worrying about user-written `sorry`s.

In the future we could have some more precise logic for sorry reporting. All the sorries in a declaration should be considered to be reported, and we should not re-report sorries in later declarations. Some elaborators use `warn.sorry` to avoid re-reporting sorries in auxiliary declarations.