feat: performance improvements for `binop%` elaborator
This PR addresses some performance issues with `HPow.hPow` elaboration caused by default instance synthesis. Now the `binop%` elaborator makes use of the fact that purely numeric expression trees with no possibility for an expected type can be synthesized with default instances. So for example, in `x ^ 1`, if the type of `x` is known, it will force the synthesis of the default instance for `1` and of the default instance for the `HPow` instance. This eager synthesis saves some overhead that is super-quadratic in the number of pending default instance problems. In the tests there are examples where elaboration takes 6-7% of before.

In more detail:
- when computing coercion points, once we successfully unify a leaf term's type with the max-type, we try synthesizing its pending variables
- leftact/rightact expression trees are now fully visited, which lets us compute whether expressions are purely numeric
- when converting leftact/rightact trees to expressions, if the operand has a known type then we try the following: (1) if the operand is acted upon by an operator of an analyzable type then we construct the tree as normal (the types are known and the instance will be synthesized immediately), (2) if the operator is purely numeric then we convert it to an expression, then fully synthesize the binary operator, and (3) otherwise fall back to not inserting any coercions and allow normal elaboration to proceed.
- we also cache `hasCoe` and `hasHomogeneousInstance` calculations

Closes #4861