chore: remove `instWPMonad` instance and test that relies upon it
The proof of the instWPMonad instance relies on the equality of
any two terms of type `IO.RealWorld`, which is only a side effect
of the current transparent definition. Ignoring the questions
around the utility of proving things about programs in `IO`, the
semantic validity of this instance in the intended model of the
IO monad is also unclear.

I tried a few things to axiomatize this instance so it could be
put into the test file to preserve the one test section that
relies on it, but I was unsuccessful; everything I attempted
caused errors. This is probably my fault.