fix: use local metavar ids in `grind`
This PR ensures that the auxiliary temporary metavariable IDs created
by the E-matching module used in `grind` are not affected by what has
been executed before invoking `grind`. The goal is to increase
`grind`’s robustness.

For example, in the E-matching module we use `Expr.quickLt` to sort
candidates. `Expr.quickLt` depends on the `Expr` hash code, which in
turn depends on metavariable IDs. Thus, before this change, the
initial next metavariable ID at the time of `grind` invocation could
affect the order in which instances were generated, and consequently
the `grind` search.